+++ server/src/main/java/ru/practicum/shareit/booking/dao/BookingRepository.java
@@ -11,31 +11,31 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface BookingRepository extends PagingAndSortingRepository<Booking, Integer> {
-    List<Booking> findByItem_IdOrderByStartDesc(int itemId);
+public interface BookingRepository extends PagingAndSortingRepository<Booking, Long> {
+    List<Booking> findByItem_IdOrderByStartDesc(long itemId);
 
-    Page<Booking> findByBooker_IdOrderByStartDesc(int userId, Pageable pageable);
+    Page<Booking> findByBooker_IdOrderByStartDesc(long userId, Pageable pageable);
 
-    Page<Booking> findByBooker_IdAndEndBeforeOrderByStartDesc(int userId, LocalDateTime now, Pageable pageable);
+    Page<Booking> findByBooker_IdAndEndBeforeOrderByStartDesc(long userId, LocalDateTime now, Pageable pageable);
 
     Page<Booking> findByBooker_IdAndEndAfterAndStartBeforeOrderByStartDesc(
-            int userId, LocalDateTime now, LocalDateTime now1, Pageable pageable);
+            long userId, LocalDateTime now, LocalDateTime now1, Pageable pageable);
 
-    Page<Booking> findByBooker_IdAndStartAfterOrderByStartDesc(int userId, LocalDateTime now, Pageable pageable);
+    Page<Booking> findByBooker_IdAndStartAfterOrderByStartDesc(long userId, LocalDateTime now, Pageable pageable);
 
-    Page<Booking> findByBooker_IdAndStatusOrderByStartDesc(int userId, BookingStatus status, Pageable pageable);
+    Page<Booking> findByBooker_IdAndStatusOrderByStartDesc(long userId, BookingStatus status, Pageable pageable);
 
-    Page<Booking> findByItem_OwnerIdOrderByStartDesc(int ownerId, Pageable pageable);
+    Page<Booking> findByItem_OwnerIdOrderByStartDesc(long ownerId, Pageable pageable);
 
     Page<Booking> findByItem_OwnerIdAndEndAfterAndStartBeforeOrderByStartDesc(
-            int ownerId, LocalDateTime now, LocalDateTime now1, Pageable pageable);
+            long ownerId, LocalDateTime now, LocalDateTime now1, Pageable pageable);
 
-    Page<Booking> findByItem_OwnerIdAndEndBeforeOrderByStartDesc(int ownerId, LocalDateTime now, Pageable pageable);
+    Page<Booking> findByItem_OwnerIdAndEndBeforeOrderByStartDesc(long ownerId, LocalDateTime now, Pageable pageable);
 
-    Page<Booking> findByItem_OwnerIdAndStartAfterOrderByStartDesc(int ownerId, LocalDateTime now, Pageable pageable);
+    Page<Booking> findByItem_OwnerIdAndStartAfterOrderByStartDesc(long ownerId, LocalDateTime now, Pageable pageable);
 
-    Page<Booking> findByItem_OwnerIdAndStatusOrderByStartDesc(int ownerId, BookingStatus waiting, Pageable pageable);
+    Page<Booking> findByItem_OwnerIdAndStatusOrderByStartDesc(long ownerId, BookingStatus waiting, Pageable pageable);
 
     List<Booking> findByItem_IdAndBooker_IdAndStatusAndEndBefore(
-            int itemId, int userId, BookingStatus status, LocalDateTime now);
+            long itemId, long userId, BookingStatus status, LocalDateTime now);
 }
