+++ server/src/main/java/ru/practicum/shareit/booking/dto/BookingSimpleDto.java
@@ -12,12 +12,12 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookingSimpleDto {
-    private int id;
+    private long id;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime start;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime end;
-    private int itemId;
-    private int bookerId;
+    private long itemId;
+    private long bookerId;
     private BookingStatus status;
 }
