+++ server/src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -12,10 +12,10 @@ import javax.validation.constraints.NotNull;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ItemDto {
-    private int id;
+    private long id;
     private String name;
     private String description;
     private Boolean available;
     private UserDto owner;
-    private Integer requestId;
+    private Long requestId;
 }
