+++ server/src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -5,13 +5,13 @@ import ru.practicum.shareit.user.dto.UserDto;
 import java.util.List;
 
 public interface UserService {
-    UserDto getUserById(int id);
+    UserDto getUserById(long id);
 
     List<UserDto> getAllUsers();
 
     UserDto createUser(UserDto userDto);
 
-    UserDto updateUser(int id, UserDto userDto);
+    UserDto updateUser(long id, UserDto userDto);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
